Nikita bg

+++ app/src/main/java/com/example/tutorsnotebook/entities/Student.kt
@@ -1,12 +1,13 @@
 package com.example.tutorsnotebook.entities
 
+//TODO: имя и фамилия с большой буквы
 class Student(
     var key: String = "key",
     var name: String = "name",
     var surname: String = "surname",
-    var studentPhone: Int = 0,
+    var studentPhone: Long = 0,
     var parentName: String = "parent name",
-    var parentPhone: Int = 0,
+    var parentPhone: Long = 0,
     var isPayed: Boolean = false,
     var avgScore: Int = 0,
     // to update average score correctly
